refactor: setupDom() 테스트 헬퍼 공통 모듈 추출 (#144)
- tests/helpers.js 신규 생성: FakeGainNode, FakeOscillator, FakeBufferSource,
  FakeAudioContext, setupDom() 공통 추출
- tests/smoke.test.js: helpers.js import으로 중복 코드 제거
- tests/unit.test.js: 동일하게 중복 코드 제거
